requirements.txt
flask
python-dotenv
Pillow
requests
gTTS
numpy
